requirements-training.txt
torchvision
braceexpand
webdataset
tqdm
wandb
scipy
torchvision
nltk
inflection
pycocoevalcap
pycocotools
scikit-learn
black
mypy
pylint
pytest
requests
